SCUMM: Encapsulate the new Doublefine SE sound code in a separate class
Also, change the Loom CDDA class so that it instantiates the CD audio
stream internally, to streamline and simplify the code in the Sound
class
